// DXC_Frontend/src/api/models/tourDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { TourImageDto } from './tourImageDto'

export interface TourDto {
  publicId?: string
  /** @nullable */
  name?: string | null
  /** @nullable */
  description?: string | null
  /** @nullable */
  highlights?: string | null
  /** @nullable */
  schedule?: string | null
  price?: number
  /** @nullable */
  priceCurrency?: string | null
  durationDays?: number
  durationNights?: number
  /** @nullable */
  departureLocation?: string | null
  maxParticipants?: number
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  /** @nullable */
  images?: TourImageDto[] | null
}